refactor: WorldModel Phase 1 — extract request handling + self-improvement
Extract 22 methods from WorldModel class into 5 standalone function modules:
- request_handling.py (223 lines): process_request + 5 type-specific handlers
- request_dispatch.py (166 lines): self-referential, introspection, conversational
- request_formatting.py (152 lines): LLM formatting, reasoning engine invocation
- self_improvement_engine.py (177 lines): autonomous improvement loop
- self_improvement_apply.py (205 lines): code improvement execution

Architecture: functions take `wm` (WorldModel instance) as first parameter.
WorldModel methods replaced with thin delegation stubs.
world_model_core.py: 8,435 → 7,339 lines (-1,096).

Phase 1 of 8 in deep God class decomposition plan.
